api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings>=2.0.0
anthropic>=0.18.1
python-multipart==0.0.9
firebase-admin==6.4.0
boto3==1.34.69
pytest==8.0.0
httpx==0.26.0
alembic==1.13.1
python-jose==3.3.0
email-validator==2.1.0.post1
passlib==1.7.4
bcrypt==4.1.2
requests==2.31.0
aiohttp==3.9.3
stripe